+++ R/dici.R
@@ -2,7 +2,6 @@
 #'
 #' Estimates of dispersal (in km^2) for the simulated mosquito population when \code{scale = "lc"}.
 #' @param input_sim matrix. \code{dynamAedes.m} compressed output matrix (\code{compressed=TRUE}).
-#' @param coords matrix. A matrix reporting the spatial coordinates of the temperature observations.
 #' @param eval_date numeric. Define the day of evaluation; it refers to the column number of the input temperature matrix. 
 #' @param breaks numeric vector. Quantile breaks, default interquartile range: \code{c(0.25,0.5,0.75)}.
 #' @param space See below for more details.
@@ -10,9 +9,16 @@
 #' @author Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, Daniele Da Re \email{daniele.dare@uclouvain.be}
 #' @export
 
-dici <- function(input_sim=NULL, coords=NULL, eval_date=NULL, breaks=c(0.25,0.5,0.75), space=FALSE) {
+dici <- function(input_sim=NULL, eval_date=NULL, breaks=c(0.25,0.5,0.75), space=FALSE) {
+		if(!is.numeric(eval_date)) {
+		stop("eval_date not defined, exiting...")
+	}	
+
+	coords = input_sim@coordinates
+	input_sim = input_sim@simulation
+
 	if( max(eval_date) > max(sapply(input_sim,length)) ) {
-		stop("eval_date > than number of simulated days...")
+		stop("eval_date > number of simulated days...")
 	}
 	if( all(unlist(lapply(input_sim, function(x) { sapply(x,length) } ))==4) ) {
 		stop("Non-dispersal model, set scale='lc' in dynamAedes.m()")