+++ R/psi_sp.R
@@ -2,14 +2,20 @@
 #'
 #' Compute the proportion of successful introductions per each cell of the grid.
 #' @param input_sim matrix. \code{dynamAedes.m} compressed output matrix (\code{compressed=TRUE}).
-#' @param coords matrix. A matrix reporting the spatial coordinates of temperature observations.
 #' @param eval_date positive integer. Define the day(s) to calculate the proportion of successful introductions which should match the column number of the temperature matrix used to inform the model.
 #' @param n.clusters positive integer. Define the number of parallel processes.
 #' @return \code{psi_sp} returns a raster with the proportion of model iterations that resulted in a viable mosquito population at a given date for a given life stage in each cell of the grid.
 #' @author Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, Daniele Da Re \email{daniele.dare@uclouvain.be}
 #' @export
 
-psi_sp = function(input_sim=NULL,coords=NULL, eval_date=NULL, n.clusters=1) {
+psi_sp = function(input_sim=NULL, eval_date=NULL, n.clusters=1) {
+	if(!is.numeric(eval_date)) {
+		stop("eval_date not defined, exiting...")
+	}	
+
+	coords = input_sim@coordinates
+	input_sim = input_sim@simulation
+	
 	if( all(unlist(lapply(input_sim, function(x) { sapply(x,length) } ))==4) ) {
 		stop("Non-spatial data, set scale='lc' or scale='rg' in dynamAedes")
 	} else {
@@ -37,7 +43,6 @@ psi_sp = function(input_sim=NULL,coords=NULL, eval_date=NULL, n.clusters=1) {
 			mylist = lapply(mylist, "[",  my_names)
   			#matrix operation
 			my.out=Reduce('+', mylist)/length(input_sim)
-  		#rasterize
 		}
 		rateList <- apply(my.out, 2, function(x) {
 			rate.sp=as.data.frame(cbind(coords, data.frame(x)))